Prevent update or delete vehicle when it has active parkings

+++ app/Http/Controllers/Api/V1/VehicleController.php
@@ -32,6 +32,12 @@ class VehicleController extends Controller
 
     public function update(StoreVehicleRequest $request, Vehicle $vehicle)
     {
+        if ($vehicle->hasActiveParkings()) {
+            return response()->json([
+                'errors' => ['general' => ['Can\'t update vehicle with active parkings. Stop active parking.']],
+            ], Response::HTTP_UNPROCESSABLE_ENTITY);
+        }
+
         $vehicle->update($request->validated());
 
         return response()->json(VehicleResource::make($vehicle), Response::HTTP_ACCEPTED);
@@ -39,6 +45,12 @@ class VehicleController extends Controller
 
     public function destroy(Vehicle $vehicle)
     {
+        if ($vehicle->hasActiveParkings()) {
+            return response()->json([
+                'errors' => ['general' => ['Can\'t remove vehicle with active parkings. Stop active parking.']],
+            ], Response::HTTP_UNPROCESSABLE_ENTITY);
+        }
+
         $vehicle->delete();
 
         return response()->noContent();
